+++ src/main/java/com/paulzzh/yuzu/init/ManosabaSounds.java
@@ -9,6 +9,8 @@ public final class ManosabaSounds {
     public static SoundEvent TITLE_MUSIC;
     public static SoundEvent BUTTON_CLICK_SUBMIT;
     public static SoundEvent BUTTON_CLICK_START_GAME;
+    public static SoundEvent SFX_SYSTEM_LOADDATA;
+    public static SoundEvent SFX_SYSTEM_STARTGAME;
 
     private ManosabaSounds() {
     }
@@ -17,6 +19,8 @@ public final class ManosabaSounds {
         TITLE_MUSIC = register("music");
         BUTTON_CLICK_SUBMIT = register("button_click_submit");
         BUTTON_CLICK_START_GAME = register("button_click_start_game");
+        SFX_SYSTEM_LOADDATA = register("sfx_system_loaddata_001");
+        SFX_SYSTEM_STARTGAME = register("sfx_system_startgame_001");
     }
 
     private static SoundEvent register(String name) {